\name{lung}
\alias{lung}
\alias{lung cancer microarray data set}


\docType{data}

\title{lung}
\description{Lung cancer gene expression data set}
\usage{
  data(lung)
  }

\format{
This data set contain 56 samples and gene expression values of 12 625 genes measured
using the Affymetrix 95av2 GeneChip. The samples comprise
20 pulmonary carcinoid samples (Carcinoid), 13 colon cancer
metastasis samples (Colon), 17 normal lung samples (Normal)
and 6 small cell lung carcinoma samples (SmallCell).
The rownames are affymetrix gene ids. 
}
%\details{}
%\value{}
\references{Bhattacharjee, A., Richards, W. G., Staunton, J., Li, C., Monti, S., Vasa, P., Ladd, C.,<br>
Beheshti, J., Bueno, R., Gillette, M., Loda, M., Weber, G., Mark, E. J., Lander,<br>
E. S., Wong, W., Johnson, B. E., Golub, T. R., Sugarbaker, D. J., and Meyerson,<br>
M. (2001). Classification of human lung carcinomas by mRNA expression profiling<br>
reveals distinct adenocarcinoma subclasses. Proceedings of the National Academy<br>
of Sciences of the United States of America.}


\source{http://www.pnas.org/content/98/24/13790/suppl/DC1
}

\keyword{datasets}