\name{stabpath}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Stability paths plot
}
\description{
The function plots the stability path of a S4VD result
}
\usage{
stabpath(res,number)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{res}{
the S4VD result  
}
  \item{number}{
the bicluster for which the stability path shall be plotted
}
}
\details{
plots the stability path the if savepath=TRUE. If the pointwise error control was used or if savepath=FALSE the final selection probabilities for the rows and the columns will be plotted.  
}
%\value{
%}
%\references{
%
%kommt noch
%}
\author{
Martin Sill
}
%\note{
%kommt noch
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%kommt noch
%}
%\examples{
%}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
